+++ CHANGES.rst
@@ -15,6 +15,7 @@ Changelog
 - Added helper methods to manipulate WF to ease applcation of workflow adaptations :
     - model.adaptations.change_transition_new_state_id to change the new_state_id of a given transition_id
     - model.adaptations.removeState that removes a state and removes transitions leading to this state and manage new initial state if necessary
+- Added workflow actions/conditions adapters for MeetingAdvice as it is already the case for Meeting/MeetingItem
 
 4.1b17 (2019-05-16)
 -------------------